moving janus cli to src folder
