+++ .gitignore
@@ -1,3 +1,3 @@
-janus-cli/docker-compose.yml
+src/janus-cli/docker-compose.yml
 ./docker/
 ./vscode/
